Lots of changes to the side view/main view engine.
- reimplemented zoom and shifting for side view
- added buttons to main view (zoom/reset)
- added shortcut/button to toggle "point to voxel" snap
- further commenting and optimization
- restricted movement of points to the box they live in
- added "ground" plane to help with orientation
- added colored axis (top left) for further orientation
- found a bug in the rtree implementation I'm using. Currently in touch with the developer to see if he can fix it. There are not many lightweight, multidimensional rtree implementations around sadly

Temporarily removed RTree.

+++ PS4k/src/com/vitco/frames/engine/data/animationdata/AnimationData.java
@@ -33,9 +33,9 @@ public class AnimationData extends AnimationDataCore implements AnimationDataInt
     }
 
     @Override
-    public int[][][] getPreviewLine() {
+    public float[][][] getPreviewLine() {
         if (isValid(previewLine[0]) && isValid(previewLine[1])) {
-            return new int[][][] {
+            return new float[][][] {
                     getPoint(previewLine[0]),
                     getPoint(previewLine[1])
             };
